+++ src/features/LandingPage/components/__tests__/ViewModeControl.test.tsx
@@ -1,9 +1,8 @@
 import { render } from "@test/testUtils";
 import { fireEvent, screen } from "@testing-library/react";
 
+import { ViewModeControl } from "@/components/viewMode/ViewModeControl";
 import { ViewModeContext } from "@/hooks/useViewMode";
-
-import { ViewModeControl } from "../ViewModeControl";
 
 describe("ViewModeControl component", () => {
   it("renders buttons for each view option", () => {
