+++ src/features/Place/components/PlaceGallery.tsx
