+++ src/features/Place/components/PlacePreviewCard.tsx
