+++ src/features/Place/components/Stats/StatDisplaySkeleton.tsx
